+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Fixes
 
+* Filetype detection if a CSV has a text/plain MIME type
 * Update the `read_txt_file` utility function to keep using `spooled_to_bytes_io_if_needed` for xml
 * Add functionality to the `read_txt_file` utility function to handle file-like object from URL
 * Remove the unused parameter `encoding` from `partition_pdf`